Get all records in a custom inspection

Get all records in a custom inspection

Usage

get_records(
  inspection_name = "Vespa-Watch",
  access_token = get_access_token(quiet = TRUE),
  get_metadata = c("id", "all")
)

Arguments

inspection_name

name of the custom inspection to return records for

access_token

access token from get_access_token()

get_metadata

Indicates whether to include metadata in the returned data. Possible values are "id" (only include observation insp_order) and "all" (include all available metadata). Default is "id".

Value

a tibble with the records from the selected inspection.

Examples

if (FALSE) get_records("Vespa-Watch") # \dontrun{}